Merge `unstable-quests` into `unstable`

+++ src/main/java/emu/grasscutter/game/managers/ResinManager.java
@@ -1,5 +1,7 @@
 package emu.grasscutter.game.managers;
 
+import static emu.grasscutter.config.Configuration.GAME_OPTIONS;
+
 import emu.grasscutter.game.inventory.GameItem;
 import emu.grasscutter.game.player.BasePlayerManager;
 import emu.grasscutter.game.player.Player;
@@ -11,12 +13,10 @@ import emu.grasscutter.server.packet.send.PacketItemAddHintNotify;
 import emu.grasscutter.server.packet.send.PacketResinChangeNotify;
 import emu.grasscutter.utils.Utils;
 
-import static emu.grasscutter.config.Configuration.GAME_OPTIONS;
-
 public class ResinManager extends BasePlayerManager {
     public static final int MAX_RESIN_BUYING_COUNT = 6;
     public static final int AMOUNT_TO_ADD = 60;
-    public static final int[] HCOIN_NUM_TO_BUY_RESIN = new int[]{50, 100, 100, 150, 200, 200};
+    public static final int[] HCOIN_NUM_TO_BUY_RESIN = new int[] {50, 100, 100, 150, 200, 200};
 
     public ResinManager(Player player) {
         super(player);
@@ -115,8 +115,9 @@ public class ResinManager extends BasePlayerManager {
         // logged off with uncapped resin and is now logging in again.
         int recharge =
                 1
-                        + ((currentTime - this.player.getNextResinRefresh())
-                                / GAME_OPTIONS.resinOptions.rechargeTime);
+                        + (int)
+                                ((currentTime - this.player.getNextResinRefresh())
+                                        / GAME_OPTIONS.resinOptions.rechargeTime);
         int newResin = Math.min(GAME_OPTIONS.resinOptions.cap, currentResin + recharge);
         int resinChange = newResin - currentResin;
 
@@ -164,7 +165,10 @@ public class ResinManager extends BasePlayerManager {
             return RetcodeOuterClass.Retcode.RET_RESIN_BOUGHT_COUNT_EXCEEDED_VALUE;
         }
 
-        var res = this.player.getInventory().payItem(201, HCOIN_NUM_TO_BUY_RESIN[this.player.getResinBuyCount()]);
+        var res =
+                this.player
+                        .getInventory()
+                        .payItem(201, HCOIN_NUM_TO_BUY_RESIN[this.player.getResinBuyCount()]);
         if (!res) {
             return RetcodeOuterClass.Retcode.RET_HCOIN_NOT_ENOUGH_VALUE;
         }
@@ -172,7 +176,8 @@ public class ResinManager extends BasePlayerManager {
         this.player.setResinBuyCount(this.player.getResinBuyCount() + 1);
         this.player.setProperty(PlayerProperty.PROP_PLAYER_WAIT_SUB_HCOIN, 0);
         this.addResin(AMOUNT_TO_ADD);
-        this.player.sendPacket(new PacketItemAddHintNotify(new GameItem(106, AMOUNT_TO_ADD), ActionReason.BuyResin));
+        this.player.sendPacket(
+                new PacketItemAddHintNotify(new GameItem(106, AMOUNT_TO_ADD), ActionReason.BuyResin));
 
         return 0;
     }
